requirements.txt

Django==4.1.7
whitenoise==6.5.0
django-recaptcha==3.0.0
django-secrets==1.2.4


